BTrack merged with the original tracker. Sort of a hybrid, way more efficient though

+++ app/src/main/cpp/tracker/CSD_detection_function.hpp
@@ -13,7 +13,7 @@ namespace reBass
     public:
         CSD_detection_function(unsigned int frameLength, unsigned int stepSize);
 
-        double process_magnitudes(
+        float process_magnitudes(
                 const std::vector<float>& magnitudes
         );
 
@@ -21,7 +21,7 @@ namespace reBass
         constexpr static double DB_RISE = 3.0;
     private:
         void whiten();
-        double broadband();
+        float broadband();
 
         unsigned int frameLength;
         unsigned int stepSize;
